fix(dashboard): extract shell footer module and tighten link tests
Move footer chrome into shell-footer.tsx and assert footer link labels stay unique.
